feat: Support loading a combined JSON file
In addition to loading a set of packages from a directory of JSON files,
also support loading such a set from a combined JSON file. This combined
file is expected to contain a top-level JSON dictionary which maps the
filename (without `.json` suffix) to the dictionary that would otherwise
be contained in that file.